Build 11: Breaking Change (#1005)
* store signature inside the change itself

* move syncing code inside vcs

* rename vcsdb to sqlitevcs

* move out generic things from vcsdb to vcs

* replace datom factory with datom writer

* hlc

* change and permanode owners 1-to-1

* implement node ID

* implement node ID tests failing

* tests pass

* fix some lint errors [skip ci]

* WIP

* refactor(sqlitegen): int to int64to avoid precision loss

* Fix lint errors

* fix tests

Former-commit-id: 4fa92d5aa695f4deaf84996e55f5bc3dc0142374

+++ backend/daemon/api/documents/v1alpha/content_graph.go
@@ -28,7 +28,7 @@ func (srv *Server) ListCitations(ctx context.Context, in *documents.ListCitation
 		return nil, err
 	}
 
-	list, err := vcssql.BacklinksListByTargetDocument(conn, docdb.IPFSBlocksID, int(in.Depth))
+	list, err := vcssql.BacklinksListByTargetDocument(conn, docdb.IPFSBlocksID, int64(in.Depth))
 	if err != nil {
 		return nil, err
 	}
